refactor: finish clean tasks naming mess
